PXC-3470: Merge codership/galera 25.3.31 -> percona/galera 3.x (Fix test failure)
https://jira.percona.com/browse/PXC-3470

***
The test `galera_3nodes.galera_join_with_cc_A` was resulting in node's local
state seqno to be greater than the currently processing CC seqno and thus was
failing with below error in the logs

    Local state seqno (9) is greater than group seqno (7): states
    diverged. Aborting to avoid potential data loss. Remove
    '<path>/grastate.dat' file and restart if you wish to continue.

Upstream fixed this issue as part of commit id 065e484144c5999709ae8fd19844,
but however was merged incorrectly causing the test to fail.

This patch makes PXC to have same behavior as of upstream i.e, actions having
seqno less than the local seqno are skipped and discarded.